Extract Prometheus module
